pyproject.toml
[build-system]
requires = [
    "setuptools>=69",
]
build-backend = "setuptools.build_meta"

[project]
dynamic = [
    "readme",
]
name = "whisperpy-diarizer"
version = "0.0.1"
license = "MIT"
license-files = ["LICENSE"]
description = "Python speaker diarization with OpenAI Whisper ASR & pyannote.audio for accurate multi‑speaker transcription & labeling."
authors = [
    { name = "Sebastian Alvarez", email = "csalvarez@ith.com.py" },
]
classifiers = [
    # complete classifier list: http://pypi.python.org/pypi?%3Aaction=list_classifiers
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Operating System :: Unix",
    "Operating System :: POSIX",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
    # uncomment if you test on these interpreters:
    # "Programming Language :: Python :: Implementation :: IronPython",
    # "Programming Language :: Python :: Implementation :: Jython",
    # "Programming Language :: Python :: Implementation :: Stackless",
    "Topic :: Utilities",
]
keywords = [
    # eg: "keyword1", "keyword2", "keyword3",
]
requires-python = ">=3.9"
dependencies = [
    "openai-whisper",
    "click",
    "torch",
    "torchaudio",
    "pyannote.audio>=3.0.0",
]
# rst = ["docutils>=0.11"]

[project.scripts]
whipy = "whisperpy_diarizer.cli:cli"

[project.urls]
"Sources" = "https://https://github.com/seb5433/whisperpy/seb5433/whisperpy"
"Changelog" = "https://https://github.com/seb5433/whisperpy/seb5433/whisperpy/blob/master/CHANGELOG.rst"
"Issue Tracker" = "https://https://github.com/seb5433/whisperpy/seb5433/whisperpy/issues"

[tool.ruff]
extend-exclude = ["static", "ci/templates"]
line-length = 140
src = ["src", "tests"]
target-version = "py39"

[tool.ruff.lint.per-file-ignores]
"ci/*" = ["S"]

[tool.ruff.lint]
ignore = [
    "RUF001", # ruff-specific rules ambiguous-unicode-character-string
    "S101", # flake8-bandit assert
    "S308", # flake8-bandit suspicious-mark-safe-usage
    "S603", # flake8-bandit subprocess-without-shell-equals-true
    "S607", # flake8-bandit start-process-with-partial-path
    "E501", # pycodestyle line-too-long
]
select = [
    "B", # flake8-bugbear
    "C4", # flake8-comprehensions
    "DTZ", # flake8-datetimez
    "E", # pycodestyle errors
    "EXE", # flake8-executable
    "F", # pyflakes
    "I", # isort
    "INT", # flake8-gettext
    "PIE", # flake8-pie
    "PLC", # pylint convention
    "PLE", # pylint errors
    "PT", # flake8-pytest-style
    "PTH", # flake8-use-pathlib
    "RSE", # flake8-raise
    "RUF", # ruff-specific rules
    "S", # flake8-bandit
    "UP", # pyupgrade
    "W", # pycodestyle warnings
]

[tool.ruff.lint.flake8-pytest-style]
fixture-parentheses = false
mark-parentheses = false

[tool.ruff.lint.isort]
forced-separate = ["conftest"]
force-single-line = true

[tool.ruff.format]
quote-style = "double"
